profile_ms/requirements.txt
appnope==0.1.0
coreapi
decorator==4.1.2
Django==2.0.1
django-cors-headers==2.1.0
django-extensions==1.9.8
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-simplejwt
djangorestframework-jwt==1.9.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Markdown==2.6.11
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
wheel==0.24.0
